+++ base/src/main/java/org/aya/resolve/context/Context.java
@@ -3,11 +3,11 @@
 package org.aya.resolve.context;
 
 import kala.collection.immutable.ImmutableSeq;
-import kala.collection.mutable.MutableList;
 import kala.control.Option;
 import org.aya.resolve.error.NameProblem;
 import org.aya.syntax.concrete.stmt.ModuleName;
-import org.aya.syntax.concrete.stmt.QualifiedID;
+import org.aya.syntax.context.Candidate;
+import org.aya.syntax.context.ContextView;
 import org.aya.syntax.ref.AnyVar;
 import org.aya.syntax.ref.GenerateKind;
 import org.aya.syntax.ref.LocalVar;
@@ -17,8 +17,6 @@ import org.aya.util.reporter.Reporter;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.nio.file.Path;
-import java.util.function.Function;
 import java.util.function.Predicate;
 
 /// > There is no "exception" in this library
@@ -27,60 +25,13 @@ import java.util.function.Predicate;
 /// if your code needs this information, feel free to change them!
 ///
 /// @author re-xyr
-public interface Context {
+public interface Context extends ContextView {
+  @Override
   @Nullable Context parent();
-  @NotNull Path underlyingFile();
-
-  default <T> @Nullable T iterate(@NotNull Function<@NotNull Context, @Nullable T> f) {
-    var p = this;
-    while (p != null) {
-      var result = f.apply(p);
-      if (result != null) return result;
-      p = p.parent();
-    }
-    return null;
-  }
-
-  /// The path of this module
-  default @NotNull ModulePath modulePath() {
-    var p = parent();
-    assert p != null;
-    return p.modulePath();
-  }
-
-  /// Getting a symbol by name {@param name}.
-  ///
-  /// @param name an id which probably unqualified
-  /// @return null if error
-  default @Nullable AnyVar get(@NotNull QualifiedID name, @NotNull Reporter reporter) {
-    return switch (name.component()) {
-      case ModuleName.ThisRef _ -> getUnqualified(name.name(), name.sourcePos(), reporter);
-      case ModuleName.Qualified qualified -> getQualified(qualified, name.name(), name.sourcePos(), reporter);
-    };
-  }
-
-  /// @see Context#get(QualifiedID, Reporter)
-  default @Nullable Option<AnyVar> getMaybe(@NotNull QualifiedID name, @NotNull Reporter reporter) {
-    return switch (name.component()) {
-      case ModuleName.ThisRef _ -> getUnqualifiedMaybe(name.name(), name.sourcePos(), reporter);
-      case ModuleName.Qualified qualified -> getQualifiedMaybe(qualified, name.name(), name.sourcePos(), reporter);
-    };
-  }
-
-  default MutableList<LocalVar> collect(@NotNull MutableList<LocalVar> container) {
-    return container;
-  }
 
   /// @return all symbols with name {@param name}
   /// @implSpec return null if not found
   @Nullable Candidate<AnyVar> getCandidateLocalMaybe(@NotNull String name, @NotNull SourcePos sourcePos);
-
-  default @Nullable Candidate<AnyVar> getCandidateMaybe(@NotNull String name, @NotNull SourcePos sourcePos) {
-    return iterate(c -> {
-      var candy = c.getCandidateLocalMaybe(name, sourcePos);
-      return candy == null || candy.isEmpty() ? null : candy;
-    });
-  }
 
   /// Trying to get a symbol by unqualified name {@param name} in `this` context.
   default @Nullable Option<AnyVar> getUnqualifiedLocalMaybe(@NotNull String name, @NotNull SourcePos sourcePos, @NotNull Reporter reporter) {
@@ -95,16 +46,6 @@ public interface Context {
     return Option.some(candy.get());
   }
 
-  /// Trying to get a symbol which can referred by unqualified name {@param name} in the whole context.
-  ///
-  /// @param name      the unqualified name
-  /// @param sourcePos the source pos for error reporting
-  /// @return null if not found, `Option.none()` if error
-  /// @see Context#getUnqualifiedLocalMaybe
-  default @Nullable Option<AnyVar> getUnqualifiedMaybe(@NotNull String name, @NotNull SourcePos sourcePos, @NotNull Reporter reporter) {
-    return iterate(c -> c.getUnqualifiedLocalMaybe(name, sourcePos, reporter));
-  }
-
   /// @return null if error
   /// @see Context#getUnqualified(String, SourcePos, Reporter)
   default @Nullable AnyVar getUnqualified(@NotNull String name, @NotNull SourcePos sourcePos, @NotNull Reporter reporter) {
@@ -114,28 +55,6 @@ public interface Context {
       return null;
     }
     return result.getOrNull();
-  }
-
-  /// Trying to get a symbol by qualified id `{modName}::{name}` in `this` context
-  ///
-  /// @return a symbol in component {@param modName}, even it is [#This]; null if not found; Option.none() if error
-  @Nullable Option<AnyVar> getQualifiedLocalMaybe(
-    @NotNull ModuleName.Qualified modName,
-    @NotNull String name,
-    @NotNull SourcePos sourcePos,
-    @NotNull Reporter reporter
-  );
-
-  /// Trying to get a symbol by qualified id `{modName}::{name}` in the whole context with {@param accessibility}.
-  ///
-  /// @see Context#getQualifiedLocalMaybe(ModuleName.Qualified, String, SourcePos, Reporter)
-  default @Nullable Option<AnyVar> getQualifiedMaybe(
-    @NotNull ModuleName.Qualified modName,
-    @NotNull String name,
-    @NotNull SourcePos sourcePos,
-    @NotNull Reporter reporter
-  ) {
-    return iterate(c -> c.getQualifiedLocalMaybe(modName, name, sourcePos, reporter));
   }
 
   /// @return null if error
@@ -154,27 +73,16 @@ public interface Context {
     return result.getOrNull();
   }
 
-  /// Trying to get a [ModuleExport] by a module {@param modName} in `this` context.
-  ///
-  /// @param modName qualified module name
-  /// @return a ModuleExport of that module; null if no such module.
-  @Nullable ModuleExport getModuleLocalMaybe(@NotNull ModuleName.Qualified modName);
-
-  /// Trying to get a [ModuleExport] by a module {@param modName} in the whole context.
-  ///
-  /// @param modName qualified module name
-  /// @return a ModuleExport of that module; null if no such module.
-  default @Nullable ModuleExport getModuleMaybe(@NotNull ModuleName.Qualified modName) {
-    return iterate(c -> c.getModuleLocalMaybe(modName));
+  @Override
+  @NotNull
+  default Context bind(@NotNull LocalVar ref, @NotNull Reporter reporter) {
+    return (Context) ContextView.super.bind(ref, reporter);
   }
 
-  default @NotNull Context bind(@NotNull LocalVar ref, @NotNull Predicate<@Nullable Candidate<AnyVar>> toWarn, @NotNull Reporter reporter) {
-    return bind(ref.name(), ref, toWarn, reporter);
-  }
-
-  default @NotNull Context bind(@NotNull LocalVar ref, @NotNull Reporter reporter) {
-    return bind(ref, var -> var instanceof Candidate.Defined<AnyVar> defined
-      && defined.get() instanceof LocalVar, reporter);
+  @Override
+  @NotNull
+  default Context bind(@NotNull LocalVar ref, @NotNull Predicate<@Nullable Candidate<AnyVar>> toWarn, @NotNull Reporter reporter) {
+    return (Context) ContextView.super.bind(ref, toWarn, reporter);
   }
 
   default @NotNull Context bind(
