+++ base/src/main/java/org/aya/resolve/context/EmptyContext.java
@@ -4,6 +4,8 @@ package org.aya.resolve.context;
 
 import kala.control.Option;
 import org.aya.syntax.concrete.stmt.ModuleName;
+import org.aya.syntax.context.Candidate;
+import org.aya.syntax.context.ModuleExport;
 import org.aya.syntax.ref.AnyVar;
 import org.aya.syntax.ref.ModulePath;
 import org.aya.util.position.SourcePos;
