+++ cli-impl/src/main/java/org/aya/cli/interactive/ReplContext.java
@@ -14,6 +14,9 @@ import kala.tuple.Tuple2;
 import org.aya.resolve.context.*;
 import org.aya.syntax.concrete.stmt.ModuleName;
 import org.aya.syntax.concrete.stmt.Stmt;
+import org.aya.resolve.context.Context;
+import org.aya.syntax.context.ModuleExport;
+import org.aya.syntax.context.ModuleSymbol;
 import org.aya.syntax.ref.AnyDefVar;
 import org.aya.syntax.ref.AnyVar;
 import org.aya.syntax.ref.DefVar;
