+++ syntax/src/main/java/org/aya/syntax/context/Candidate.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2025 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.resolve.context;
+package org.aya.syntax.context;
 
 import kala.collection.CollectionView;
 import kala.collection.SeqView;
